New Commands Update
Added a new command for fake damage and even combat alerts depending of the entity type of the structure selected and the ability to define a maximum ammount of entities. Added a new command for random safe teleportation in a pre definable radius. Added a new Command for triggering attacks with a random selection of biter and spitters of all sizes and a definable maximum amount.