Fixes & Improvements.
Fixed: Mod accidentally evaluated against settings Prototypes rather then the actual setting.
Added: Error Failsafe to avoid crashes when the mod is unable to resolve channel name or player name.
Removed: Unecessary Settings Caching previously done by the Mod.
Fixed: Mistake in the Mods Logic sometimes lead to attempting the use of Data before it became available.
Fixed: Typo in the Attack Ai Command Structure.
Fixed: Mod used a LuaPlayer reference instead of a LuaCharacter reference for "destination_entity" and "target" in the Attack Ai Commands.
Changed: Mod used to select random Bugs on the Map for the Ai Attack commands which lead to a trickle rather then a Wave, now it Spawns Bugs for the Attack. Will always attempt to Spawn near the closest Bug Spawner.
Changed: Mod had an unstable behaviour when determing which entities to raise an Alert. Now it should be much more stable.
Fixed: Mod used to use a LuaPlayer reference instead of a LuaCharacter for "find_non_colliding_position" in the Teleport victim to random nearby tile Event. Mod will now select a random angle in which direction to move the Player.